Move engine logic out of models
The engine will be separate and the models will be simple value objects.

Left the engine spec as pending for now.
